super class for gas mining facilities.
